Rename authapi package to auth; add empty admin module
